:recycle: [Refactor] ChatGpt 요약 요청 본문 구체화
